feat(rdaspen/ras): RAS intr handling for PC CPU
This introduces and enables the RAS framework and enables
the RAS for CPU handling.

This commit ensures that RAS settings are initialized on
all CPU cores during the firmware boot stage. Previously,
the initialization for RAS processing was only done on
the primary CPU core.

This also introduces a custom bl31_platform_setup to
allow RAS specific intialisations.

Upstream-Status: Pending
Signed-off-by: Ahmed Azeem <ahmed.azeem@arm.com>
Signed-off-by: Ahmed Tiba <ahmed.tiba@arm.com>