fix(rdaspen): fix TrustZone extension check
Checking if TrustZone extension is supported relied on MHUv3
driver. This driver is currently being used for AP<->RSE communication
and not AP<->SI CL0, resulting in the incorrect MHUv3 device.

Also, the previous implementation applied its changes in the BL2,
but this check needs to be done in BL31.

Upstream-Status: Pending
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>