fix(rdaspen): disable access to system counter
For some reason on RTL testbench, an exception occurs whenever
the system counter is accessed from TF-A. This access is not
really required on Aspen because SI CL0 already enables the counter,
but still the access should work fine. Disable the write until
the problem is discovered.

Upstream-Status: Inappropriate [RTL workaround]
Signed-off-by: Ziad Elhanafy <ziad.elhanafy@arm.com>