refac(rdaspen) ras: remove mhu sync
This removes the mhu sync between AP and SI
the reason is to have all inband errors to be
handled on AP according to a CPU RAS handling
design update.

Inband errors only consist of:
- Corrected Errors
- Deferred Errors

It also fixes a compilation issue with cper
misaligned pointer passed to print_guid and now
a copy of the struct is passed instead.

Upstream-Status: Pending
Signed-off-by: Ahmed Azeem <ahmed.azeem@arm.com>